Add feature requests to BUGS.txt
A multicall binary will eliminate a lot of the header bloat from the excessive number of libraries.
The GPL3 code could be backported with permission to relicense and save some reinventing the wheel.
Some of the code is unnecessarily complex, so fix that where practical.  We don't _need_ backwards compatibility until we mark a major release, but don't break it just because.